Add progress bar streak colors
The challenging part here was getting the gradient to not change, but rather be
static and revealed gradually. I played around with a few options, initially
tried a mask but `MaskedView` only works on mobile (not web). Instead tried the
`onLayout` approach which meant the gradient could have an absolute width and
then its container have `overflow: hidden` to mask it.