Downgrade Node.js to v21
This fixes an issue with Expo where it was logging warnings:

```
Warning: To load an ES module, set "type": "module" in the package.json or use the .mjs extension.
```

It would output one line for each `_layout.tsx` in the project. The issue only
seems to have started in Node.js 22. Another work-around is to convert these
files to CommonJS with `require()` and `module.exports =`, but that will create
more refactoring burden later. There's a reddit thread about it too
https://www.reddit.com/r/expo/comments/1cvhzlr/constantly_getting_a_warning_to_load_an_es_module/

Instead for now downgrading to v21.

Also added instructions for how to change the Node.js version, and reported a
bug to Moon for it not respecting ancestor `.prototools` for the Node.js
version.
https://discord.com/channels/974160221452763146/974160221452763149/1243515551326933002