Use npx instead of Yarn dependencies for CLI tools
The yarn dependencies were super slow in CI because they had to compile native
code. Most of the time they're not even needed until the end, and doing it
up-front slows down the dev loop.

Future task is to setup renovate to keep the versions updated. Unfortunately
using npx means the transitive versions are not pinned, but I prefer the speed
win.